API Reference

MPIHaloArrays.AbstractParallelTopology — Type

An abstract AbstractParallelTopology type that is extended by either a CartesianTopology or GraphTopology (future)

MPIHaloArrays.CartesianTopology — Type

CartesianTopology

The CartesianTopology type holds neighbor information, current rank, etc.

Fields

comm: MPI commicator object

nprocs: Number of total processors (global)

rank: Current rank

coords: Coordinates in the global space, i.e. (0,1,1)

global_dims: Dimensions of the global domain, i.e. (4,4) is a 4x4 global domain

isperiodic: Vector{Bool}; Perodicity of each dimension, i.e. (false, true, true) means y and z are periodic

neighbors: OffsetArray{Int}; Neighbor ranks (including corners), indexed as [[ilo, center, ihi], i, j, k]

MPIHaloArrays.CartesianTopology — Method

Create CartesianTopology only with the vector of boundary periodicity given. This finds the optimal sub-domain ordering for the user.

MPIHaloArrays.CartesianTopology — Method

Create a CartesianTopology type that holds neighbor information, current rank, etc.

Arguments

dims: Dimensions of the domain in each direction, e.g. [4,3] means a total of 12 procs, with 4 in x and 3 in y

periodicity: Vector of bools to set if the domain is periodic along a specific dimension

Example

# Create a topology of 4x4 with periodic boundaries in both directions
P = CartesianTopology([4,4], [true, true])

MPIHaloArrays.MPIHaloArray — Type

MPIHaloArray

Fields

data: AbstractArray{T,N} - contains the local data on the current rank

partitioning: partitioning datatype

comm: MPI communicator

window: MPI window

neighbor_ranks : Vector{Int} - IDs of the neighboring arrays/MPI procs

coords : Vector{Int} - Coordinates in the global MPI space

rank: Current MPI rank

MPIHaloArrays.MPIHaloArray — Method

MPIHaloArray constructor

Arguments

A: AbstractArray{T,N}

topo: Parallel topology type, e.g. CartesianTopology

nhalo: Number of halo cells

Keyword Arguments

do_corners: [true] Exchange corner halo regions

com_model: [p2p] Communication model, e.g. :p2p is point-to-point (Isend, Irecv), :rma is onesided (Get,Put), :shared is MPI's shared memory model

MPIHaloArrays.coord_to_rank — Method

Helper function to find rank based on coordinates

MPIHaloArrays.denominators1 — Method

Return all common denominators of n

MPIHaloArrays.domainview — Method

Return a view of the domain data within the A::MPIHaloArray

MPIHaloArrays.gatherglobal — Method

Gather all MPIHaloArrays onto the root MPI rank and stitch together. This will ignore halo region data and create a Array that represents the global state.

Arguments

A: MPIHaloArray

root: MPI rank to gather A to

halo_dims: Tuple of the dimensions that halo exchanges occur on (not fully working yet)

MPIHaloArrays.get_dims — Method

Get the dimensions of each chunk

MPIHaloArrays.get_subdomain_dimension_sizes — Method

get_subdomain_dimension_sizes(A, tile_dims, A_halo_dims)

Get the size along each dimension in (i,j,k) of the subdomain, based on a given array A. The tile_dims is the shape of the global domain, e.g. (4,2) means 4 tiles or subdomains in i and 2 in j. A_halo_dims is the tuple of which dimensions the halo exchanges take place on, e.g. (2,3).

Example

A = rand(4,200,100); dims=(2,3), tile_dims=(4,2)
get_subdomain_dimension_sizes(A, tile_dims, dims) # [[i][j]] --> [[50,50,50,50],[100,100]]

MPIHaloArrays.globalmax — Method

Perform a global maximum operation

Arguments

A: MPIHaloArray to perform the operation on

broadcast: true/false - broadcast to all MPI ranks [default is false]

root: If broadcast is false, which MPI rank to reduce to

MPIHaloArrays.globalmin — Method

Perform a global minimum operation

Arguments

A: MPIHaloArray to perform the operation on

broadcast: true/false - broadcast to all MPI ranks [default is false]

root: If broadcast is false, which MPI rank to reduce to

MPIHaloArrays.globalsize — Method

Find the global dims of based on the list of local MPIHaloArray sizes

MPIHaloArrays.globalsum — Method

Perform a global sum operation

Arguments

A: MPIHaloArray to perform the operation on

broadcast: true/false - broadcast to all MPI ranks [default is false]

root: If broadcast is false, which MPI rank to reduce to

MPIHaloArrays.hi_indices — Method

Helper functions to get the high side halo and domain starting/ending indices

MPIHaloArrays.ihi_neighbor — Method

Neighbor rank in the i+1 direction

MPIHaloArrays.ilo_neighbor — Method

Neighbor rank in the i-1 direction

MPIHaloArrays.jhi_neighbor — Method

Neighbor rank in the j+1 direction

MPIHaloArrays.jlo_neighbor — Method

Neighbor rank in the j-1 direction

MPIHaloArrays.khi_neighbor — Method

Neighbor rank in the k+1 direction

MPIHaloArrays.klo_neighbor — Method

Neighbor rank in the k-1 direction

MPIHaloArrays.lo_indices — Method

Helper functions to get the low side halo and domain starting/ending indices

Arguments

field: Array

dim: Dimension to check indices on

nhalo: Number of halo entries

MPIHaloArrays.match_tile_halo_dim_sizes — Method

match_tile_halo_dim_sizes(tile_dims, halo_dims)

Ensure that the tile dimension tuple is the same length as the halo dim tuple. If not, then pad with ones.

MPIHaloArrays.neighbor — Method

neighbor(p::CartesianTopology, i_offset::Int, j_offset::Int, k_offset::Int)

Find the neighbor rank based on the offesets in (i,j,k). This follows the traditional array index convention rather than MPI's version, so an i_offset=1 will shift up in the array indexing.

Arguments

p : CartesianTopology type

i_offset: Offset in the i direction

j_offset: Offset in the j direction

k_offset: Offset in the k direction

Example:

# Makes a 4x4 domain with periodic boundaries in both dimensions
P = CartesianTopology([4,4], [true, true])

# Find the ihi neighbor
ihi = neighbor(P,+1,0,0)

# Find the upper ihi corner neighbor (ihi and jhi side)
ihijhi_corner = neighbor(P,+1,+1,0)

MPIHaloArrays.num_2d_tiles1 — Method

Returns the optimal number of tiles in (i,j) given total number of tiles n

MPIHaloArrays.num_3d_tiles1 — Method

Returns the optimal number of tiles in (i,j,k) given total number of tiles n

MPIHaloArrays.offset_coord_to_rank — Method

Helper function to find rank based on 3D offsets

MPIHaloArrays.offset_coord_to_rank — Method

Helper function to find rank based on 2D offsets

MPIHaloArrays.scatterglobal — Method

Partition the array A on the rank root into chunks based on the given parallel toplogy. The array data in A does not have halo regions. The MPIHaloArray constructor adds the halo regions. This returns a MPIHaloArray

Arguments

A: Global array to be split up into chunks and sent to all ranks. This does not include halo cells

root: MPI rank that A lives on

nhalo: Number of halo cells to create

halo_dims: Tuple of the dimensions that halo exchanges occur on (not fully working yet)

MPIHaloArrays.split_count — Method

split_count(N::Integer, n::Integer)

Return a vector of n integers which are approximately equally sized and sum to N.

MPIHaloArrays.sync_edges_rma! — Method

Sync the edges of the array A with it's neighbors