API Reference

MPIHaloArrays.CartesianTopology — Type

CartesianTopology

The CartesianTopology type holds neighbor information, current rank, etc.

Fields

comm: MPI commicator object

nprocs: Number of total processors (global)

rank: Current rank

coords: Coordinates in the global space, i.e. (0,1,1)

global_dims: Dimensions of the global domain, i.e. (4,4) is a 4x4 global domain

isperiodic: Vector{Bool}; Perodicity of each dimension, i.e. (false, true, true) means y and z are periodic

neighbors: OffsetArray{Int}; Neighbor ranks (including corners), indexed as [[ilo, center, ihi], i, j, k]

MPIHaloArrays.CartesianTopology — Method

Create a CartesianTopology type that holds neighbor information, current rank, etc.

Arguments

dims: Dimensions of the domain in each direction, e.g. [4,3] means a total of 12 procs, with 4 in x and 3 in y

periodicity: Vector of bools to set if the domain is periodic along a specific dimension

Example

# Create a topology of 4x4 with periodic boundaries in both directions
P = CartesianTopology([4,4], [true, true])

MPIHaloArrays.MPIHaloArray — Type

MPIHaloArray

Fields

data: AbstractArray{T,N} - contains the local data on the current rank

partitioning: partitioning datatype

comm: MPI communicator

window: MPI window

neighbor_ranks : Vector{Int} - IDs of the neighboring arrays/MPI procs

coords : Vector{Int} - Coordinates in the global MPI space

rank: Current MPI rank

MPIHaloArrays.MPIHaloArray — Method

MPIHaloArray constructor

Arguments

A: AbstractArray{T,N}

topo: Parallel topology type, e.g. CartesianTopology

nhalo: Number of halo cells

Keyword Arguments

do_corners: [true] Exchange corner halo regions

com_model: [p2p] Communication model, e.g. :p2p is point-to-point (Isend, Irecv), :rma is onesided (Get,Put), :shared is MPI's shared memory model

MPIHaloArrays.ParallelTopology — Type

An abstract ParallelTopology type that is extended by either a CartesianTopology or GraphTopology (future)

MPIHaloArrays.coord_to_rank — Method

Helper function to find rank based on coordinates

MPIHaloArrays.denominators — Method

Return all common denominators of n

MPIHaloArrays.gatherglobal — Method

Gather all MPIHaloArrays onto the root MPI rank and stitch together. This will ignore halo region data and create a Array that represents the global state.

Arguments

A: MPIHaloArray

root: MPI rank to gather A to

MPIHaloArrays.get_dims — Method

Get the dimensions of each chunk

MPIHaloArrays.globalsize — Method

Find the global dims of based on the list of local MPIHaloArray sizes

MPIHaloArrays.hi_indices — Method

Helper functions to get the high side halo and domain starting/ending indices

MPIHaloArrays.ihi_neighbor — Method

Neighbor rank in the i+1 direction

MPIHaloArrays.ilo_neighbor — Method

Neighbor rank in the i-1 direction

MPIHaloArrays.jhi_neighbor — Method

Neighbor rank in the j+1 direction

MPIHaloArrays.jlo_neighbor — Method

Neighbor rank in the j-1 direction

MPIHaloArrays.khi_neighbor — Method

Neighbor rank in the k+1 direction

MPIHaloArrays.klo_neighbor — Method

Neighbor rank in the k-1 direction

MPIHaloArrays.lo_indices — Method

Helper functions to get the low side halo and domain starting/ending indices

Arguments

field: Array

dim: Dimension to check indices on

nhalo: Number of halo entries

MPIHaloArrays.neighbor — Method

neighbor(p::CartesianTopology, i_offset::Int, j_offset::Int, k_offset::Int)

Find the neighbor rank based on the offesets in (i,j,k). This follows the traditional array index convention rather than MPI's version, so an i_offset=1 will shift up in the array indexing.

Arguments

p : CartesianTopology type

i_offset: Offset in the i direction

j_offset: Offset in the j direction

k_offset: Offset in the k direction

Example:

# Makes a 4x4 domain with periodic boundaries in both dimensions
P = CartesianTopology([4,4], [true, true])

# Find the ihi neighbor
ihi = neighbor(P,+1,0,0)

# Find the upper ihi corner neighbor (ihi and jhi side)
ihijhi_corner = neighbor(P,+1,+1,0)

MPIHaloArrays.num_2d_tiles — Method

Returns the optimal number of tiles in (i,j) given total number of tiles n

MPIHaloArrays.num_3d_tiles — Method

Returns the optimal number of tiles in (i,j,k) given total number of tiles n

MPIHaloArrays.offset_coord_to_rank — Method

Helper function to find rank based on 3D offsets

MPIHaloArrays.offset_coord_to_rank — Method

Helper function to find rank based on 2D offsets

MPIHaloArrays.scatterglobal — Method

Partition the array A on the rank root into chunks based on the given parallel toplogy. The array data in A does not have halo regions. The MPIHaloArray constructor adds the halo regions. This returns a MPIHaloArray

Arguments

A: Global array to be split up into chunks and sent to all ranks. This does not include halo cells

root: MPI rank that A lives on

nhalo: Number of halo cells to create

MPIHaloArrays.sync_edges_rma! — Method

Sync the edges of the array A with it's neighbors

MPIHaloArrays.tile_id_to_ij — Method

Given tile id in a 1D layout, returns the corresponding tile indices in a 2D layout

MPIHaloArrays.tile_id_to_ijk — Method

Given tile id in a 1D layout, returns the corresponding tile indices in a 3D layout

MPIHaloArrays.tile_indices_1d — Method

Given an input I dimensions of the total computational domain, returns an array of start and end indices [ilo,ihi]

MPIHaloArrays.tile_indices_2d — Method

Given an input (I,J) dimensions of the total computational domain, returns an array of start and end indices [ilo,ihi,jlo,jhi]

MPIHaloArrays.tile_indices_3d — Method

Given an input (I,J,K) dimensions of the total computational domain, returns an array of start and end indices [ilo,ihi,jlo,jhi,klo,khi]